requirements.txt
# PASIS - Physical AI Strategic Intelligence System
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
tenacity>=9.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Data Collection
arxiv>=2.1.0
requests>=2.32.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Web Dashboard
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0

# Scheduler
apscheduler>=3.10.4

# Utilities
python-dateutil>=2.9.0
